Update Get-ADAuditData.ps1
Added additional spaces between queries. Implemented fix for Server 2019 and above where Get-ADOrganizationUnit appears to time out. Added check for file length gt 260 and notifies person running the script of the problem. Added Try/catch block so that if the Confidentiality Bit does not exists it communicates that a little better.